Делает ДЗ №5.15
